WIP: refactor(common): move everything from common.js to plugin.utils
